python/requirements.txt
mysql-connector-python
kafka-python
pymemcache
elasticsearch
